+++ src/main/groovy/io/xh/hoist/exception/NotAuthorizedException.groovy
@@ -10,21 +10,15 @@ package io.xh.hoist.exception
 /**
  * Exception for use when the authenticated user does not have access to the resource in question.
  *
- * This exception is thrown by Hoist security - see io.xh.hoist.security.Access.
- * This exception, or subclasses of it, may be thrown directly by applications as well.
- * Instances of this exception will be associated with HttpStatus 404 ('Unauthorized') when sent to client.
+ * This exception is thrown by Hoist's AccessInterceptor class if the user does not have roles
+ * required by a controller's @Access annotation. Applications may also throw this exception, or
+ * subclasses of it, directly in response to requests they cannot fulfill due to auth-related
+ * constraints.
+ *
+ * Instances of this exception will be sent to clients with HttpStatus 403 ('Forbidden').
  */
-class NotAuthorizedException extends RuntimeException {
+class NotAuthorizedException extends RuntimeException implements RoutineException {
     NotAuthorizedException(String s = 'Not Authorized') {
-        super(s)
-    }
-}
-
-/**
- * Exception for use when the user initiating a client request does not match the session user.
- */
-class SessionMismatchException extends RuntimeException {
-    SessionMismatchException(String s = 'Client username does not match current session user.') {
         super(s)
     }
 }