+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 * A `RoutineException` interface has been added. Implement this interface to mark any exception that
   is a part of normal operations and should not necessarily be logged on the server as an error.
 * The `DataNotAvailableException` has been added. This class implements `RoutineException` and is
-  intended to be thrown
+  intended to be thrown when requested data is not currently available due to normal, expected
+  business conditions (e.g. the business day has just rolled and new data is not yet ready).
 
 ### ⚙️ Technical
 
@@ -280,8 +281,8 @@ ALTER TABLE xh_preference ALTER COLUMN group_name varchar(255) NOT NULL
 
 
 * ClientError tracking gets a `userAlerted` flag to record whether or not the user was shown a
-  pop-up dialog (vs. an error being reported quietly in the background).<br><br> ⚠️ **Note**
-  schema update required:
+  pop-up dialog (vs. an error being reported quietly in the background).<br><br> ⚠️ **Note** schema
+  update required:
 
 ```sql
 -- SQL Server